requirements.txt
Flask
requests
bs4
python-dotenv
docker